refactor(harness): integrate AgentSessionRuntime into createRunner
